refactor: remove method
